feat: reading ambasador data from out browserdb
