+++ client/src/app/layout.tsx
@@ -1,12 +1,23 @@
+'use client';
+
+import { AppProvider } from '@/context/AppContext';
+import { MeshProvider } from '@meshsdk/react';
 import type { Metadata } from 'next';
 import './app.css';
 
-export const metadata: Metadata = {
-  title: 'Cardano Ambassador Tool',
-  description: 'Discover the passionate individuals shaping the Cardano ecosystem. From developers and educators to community organizers, these ambassadors are driving innovation, connection, and real-world impact.',
-  keywords: ['Cardano', 'Ambassador', 'Blockchain', 'Cryptocurrency', 'Community'],
-  authors: [{ name: 'Cardano Ambassador Team' }],
-};
+// export const metadata: Metadata = {
+//   title: 'Cardano Ambassador Tool',
+//   description:
+//     'Discover the passionate individuals shaping the Cardano ecosystem. From developers and educators to community organizers, these ambassadors are driving innovation, connection, and real-world impact.',
+//   keywords: [
+//     'Cardano',
+//     'Ambassador',
+//     'Blockchain',
+//     'Cryptocurrency',
+//     'Community',
+//   ],
+//   authors: [{ name: 'Cardano Ambassador Team' }],
+// };
 
 export default function RootLayout({
   children,
@@ -47,7 +58,9 @@ export default function RootLayout({
         />
       </head>
       <body>
-        {children}
+        <MeshProvider>
+          <AppProvider>{children}</AppProvider>
+        </MeshProvider>
       </body>
     </html>
   );
