--- client/src/data/mockAmbassadors.ts
@@ -1,693 +1,0 @@
-export interface Ambassador {
-  id: number;
-  name: string;
-  country: string;
-  avatar?: string; // Optional - if not provided, use initials
-  isFollowing: boolean;
-  joinDate?: string,
-  contributions?: number,
-  lastActive?: string,
-  status?: 'active' | 'inactive';
-  transactionHash?: string;
-}
-
-export const mockAmbassadors: Ambassador[] = [
-  // Original 15 ambassadors
-  {
-    id: 1,
-    name: "Augustine Franchella",
-    country: "Argentina",
-    avatar: "/avatars/augustine.jpg",
-    isFollowing: false,
-    joinDate: "2022-03-15",
-    contributions: 42,
-    lastActive: "2023-06-10",
-    status: "active",
-    transactionHash: "0x4a3b2c1d0e9f8a7b6c5d4e3f2a1b0c9d8e7f6a"
-  },
-  {
-    id: 2,
-    name: "Alexandru D.",
-    country: "Romania",
-    isFollowing: true,
-    joinDate: "2022-01-12",
-    contributions: 28,
-    lastActive: "2024-08-05",
-    status: "active",
-    transactionHash: "0x5b4c3d2e1f0a9b8c7d6e5f4a3b2c1d0e9f8a7b"
-  },
-  {
-    id: 3,
-    name: "Andreas Sosilo",
-    country: "Indonesia",
-    isFollowing: false,
-    joinDate: "2021-11-20",
-    contributions: 73,
-    lastActive: "2024-07-22",
-    status: "active",
-    transactionHash: "0x6c5d4e3f2a1b0c9d8e7f6a5b4c3d2e1f0a9b8c"
-  },
-  {
-    id: 4,
-    name: "Beatrice Anihiri",
-    country: "Norway",
-    isFollowing: false,
-    joinDate: "2023-02-08",
-    contributions: 19,
-    lastActive: "2024-06-15",
-    status: "inactive",
-    transactionHash: "0x7d6e5f4a3b2c1d0e9f8a7b6c5d4e3f2a1b0c9d"
-  },
-  {
-    id: 5,
-    name: "Benjamin Baani",
-    country: "Ghana",
-    isFollowing: true,
-    joinDate: "2022-07-03",
-    contributions: 56,
-    lastActive: "2024-08-09",
-    status: "active",
-    transactionHash: "0x8e7f6a5b4c3d2e1f0a9b8c7d6e5f4a3b2c1d0e"
-  },
-  {
-    id: 6,
-    name: "Chris Kulas",
-    country: "Germany",
-    isFollowing: false,
-    joinDate: "2021-09-14",
-    contributions: 91,
-    lastActive: "2024-05-30",
-    status: "active",
-    transactionHash: "0x9f8a7b6c5d4e3f2a1b0c9d8e7f6a5b4c3d2e1f"
-  },
-  {
-    id: 7,
-    name: "Dan Baruka",
-    country: "DRC",
-    isFollowing: false,
-    joinDate: "2023-04-22",
-    contributions: 15,
-    lastActive: "2024-03-18",
-    status: "inactive",
-    transactionHash: "0xa0b9c8d7e6f5a4b3c2d1e0f9a8b7c6d5e4f3a2"
-  },
-  {
-    id: 8,
-    name: "Daniel Davide Lezu",
-    country: "Scotland",
-    isFollowing: false,
-    joinDate: "2022-10-05",
-    contributions: 38,
-    lastActive: "2024-07-12",
-    status: "active",
-    transactionHash: "0xb1c2d3e4f5a6b7c8d9e0f1a2b3c4d5e6f7a8b9"
-  },
-  {
-    id: 9,
-    name: "Daniel Lynch",
-    country: "United States",
-    isFollowing: false,
-    joinDate: "2021-12-11",
-    contributions: 67,
-    lastActive: "2024-08-02",
-    status: "active",
-    transactionHash: "0xc2d3e4f5a6b7c8d9e0f1a2b3c4d5e6f7a8b9c0"
-  },
-  {
-    id: 10,
-    name: "Darlington Wleh",
-    country: "United States",
-    isFollowing: false,
-    joinDate: "2023-01-19",
-    contributions: 24,
-    lastActive: "2024-04-28",
-    status: "inactive",
-    transactionHash: "0xd3e4f5a6b7c8d9e0f1a2b3c4d5e6f7a8b9c0d1"
-  },
-  {
-    id: 11,
-    name: "Denis Kalinin",
-    country: "Kazakhstan",
-    isFollowing: false,
-    joinDate: "2022-05-16",
-    contributions: 49,
-    lastActive: "2024-08-07",
-    status: "active",
-    transactionHash: "0xe4f5a6b7c8d9e0f1a2b3c4d5e6f7a8b9c0d1e2"
-  },
-  {
-    id: 12,
-    name: "Diego Torres Borda",
-    country: "Argentina",
-    isFollowing: false,
-    joinDate: "2021-08-09",
-    contributions: 83,
-    lastActive: "2024-06-25",
-    status: "active",
-    transactionHash: "0xf5a6b7c8d9e0f1a2b3c4d5e6f7a8b9c0d1e2f3"
-  },
-  {
-    id: 13,
-    name: "Dmitry",
-    country: "Poland",
-    isFollowing: false,
-    joinDate: "2023-03-27",
-    contributions: 31,
-    lastActive: "2024-02-14",
-    status: "inactive",
-    transactionHash: "0x0a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8"
-  },
-  {
-    id: 14,
-    name: "Edickson Rafael",
-    country: "Venezuela",
-    isFollowing: false,
-    joinDate: "2022-11-08",
-    contributions: 45,
-    lastActive: "2024-07-19",
-    status: "active",
-    transactionHash: "0x1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
-  },
-  {
-    id: 15,
-    name: "Efe Safak Senguul",
-    country: "Netherlands",
-    isFollowing: false,
-    joinDate: "2021-10-23",
-    contributions: 72,
-    lastActive: "2024-08-01",
-    status: "active",
-    transactionHash: "0x2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0"
-  },
-  {
-    id: 16,
-    name: "Elena Kowalski",
-    country: "Poland",
-    avatar: "/avatars/elena.jpg",
-    isFollowing: true,
-    joinDate: "2022-06-14",
-    contributions: 58,
-    lastActive: "2024-08-08",
-    status: "active",
-    transactionHash: "0x3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1"
-  },
-  {
-    id: 17,
-    name: "Francesco Romano",
-    country: "Italy",
-    isFollowing: false,
-    joinDate: "2023-05-11",
-    contributions: 22,
-    lastActive: "2024-05-16",
-    status: "inactive",
-    transactionHash: "0x4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2"
-  },
-  {
-    id: 18,
-    name: "Gabriela Silva",
-    country: "Brazil",
-    avatar: "/avatars/gabriela.jpg",
-    isFollowing: true,
-    joinDate: "2021-07-07",
-    contributions: 89,
-    lastActive: "2024-08-11",
-    status: "active",
-    transactionHash: "0x5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3"
-  },
-  {
-    id: 19,
-    name: "Hassan Al-Rashid",
-    country: "UAE",
-    isFollowing: false,
-    joinDate: "2022-09-03",
-    contributions: 34,
-    lastActive: "2024-06-08",
-    status: "active",
-    transactionHash: "0x6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4"
-  },
-  {
-    id: 20,
-    name: "Isabella Chen",
-    country: "Singapore",
-    isFollowing: false,
-    joinDate: "2023-08-15",
-    contributions: 17,
-    lastActive: "2024-03-22",
-    status: "inactive",
-    transactionHash: "0x7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5"
-  },
-  {
-    id: 21,
-    name: "Jacques Dubois",
-    country: "France",
-    avatar: "/avatars/jacques.jpg",
-    isFollowing: true,
-    joinDate: "2021-06-18",
-    contributions: 95,
-    lastActive: "2024-08-06",
-    status: "active",
-    transactionHash: "0x8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6"
-  },
-  {
-    id: 22,
-    name: "Kenji Nakamura",
-    country: "Japan",
-    isFollowing: false,
-    joinDate: "2022-12-02",
-    contributions: 41,
-    lastActive: "2024-07-14",
-    status: "active",
-    transactionHash: "0x9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7"
-  },
-  {
-    id: 23,
-    name: "Liam O'Sullivan",
-    country: "Ireland",
-    isFollowing: false,
-    joinDate: "2023-07-09",
-    contributions: 12,
-    lastActive: "2024-04-11",
-    status: "inactive",
-    transactionHash: "0x0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8"
-  },
-  {
-    id: 24,
-    name: "Maria Gonzalez",
-    country: "Spain",
-    avatar: "/avatars/maria.jpg",
-    isFollowing: true,
-    joinDate: "2021-05-25",
-    contributions: 104,
-    lastActive: "2024-08-10",
-    status: "active",
-    transactionHash: "0x1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9"
-  },
-  {
-    id: 25,
-    name: "Nkiru Okafor",
-    country: "Nigeria",
-    isFollowing: false,
-    joinDate: "2022-04-17",
-    contributions: 63,
-    lastActive: "2024-07-28",
-    status: "active",
-    transactionHash: "0x2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0"
-  },
-  {
-    id: 26,
-    name: "Oliver Smith",
-    country: "United Kingdom",
-    isFollowing: false,
-    joinDate: "2023-09-06",
-    contributions: 8,
-    lastActive: "2024-02-05",
-    status: "inactive",
-    transactionHash: "0x3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1"
-  },
-  {
-    id: 27,
-    name: "Priya Sharma",
-    country: "India",
-    avatar: "/avatars/priya.jpg",
-    isFollowing: true,
-    joinDate: "2021-04-13",
-    contributions: 118,
-    lastActive: "2024-08-12",
-    status: "active",
-    transactionHash: "0x4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2"
-  },
-  {
-    id: 28,
-    name: "Raj Patel",
-    country: "India",
-    isFollowing: false,
-    joinDate: "2022-08-21",
-    contributions: 52,
-    lastActive: "2024-06-19",
-    status: "active",
-    transactionHash: "0x5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2c3"
-  },
-  {
-    id: 29,
-    name: "Sofia Andersson",
-    country: "Sweden",
-    isFollowing: false,
-    joinDate: "2023-10-14",
-    contributions: 29,
-    lastActive: "2024-05-08",
-    status: "inactive",
-    transactionHash: "0x6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2c3d4"
-  },
-  {
-    id: 30,
-    name: "Thomas Mueller",
-    country: "Germany",
-    avatar: "/avatars/thomas.jpg",
-    isFollowing: true,
-    joinDate: "2021-03-30",
-    contributions: 126,
-    lastActive: "2024-08-11",
-    status: "active",
-    transactionHash: "0x7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2c3d4e5"
-  },
-  {
-    id: 31,
-    name: "Yuki Tanaka",
-    country: "Japan",
-    isFollowing: false,
-    joinDate: "2022-02-24",
-    contributions: 76,
-    lastActive: "2024-07-03",
-    status: "active",
-    transactionHash: "0x8a9b0c1d2e3f4a5b6c7d8e9f0a1b2c3d4e5f6"
-  },
-  {
-    id: 32,
-    name: "Zara Al-Mansouri",
-    country: "UAE",
-    isFollowing: false,
-    joinDate: "2023-11-28",
-    contributions: 6,
-    lastActive: "2024-01-15",
-    status: "inactive",
-    transactionHash: "0x9b0c1d2e3f4a5b6c7d8e9f0a1b2c3d4e5f6a7"
-  },
-  {
-    id: 33,
-    name: "Adrian Novak",
-    country: "Czech Republic",
-    avatar: "/avatars/adrian.jpg",
-    isFollowing: true,
-    joinDate: "2021-12-07",
-    contributions: 84,
-    lastActive: "2024-08-04",
-    status: "active",
-    transactionHash: "0x0c1d2e3f4a5b6c7d8e9f0a1b2c3d4e5f6a7b8"
-  },
-  {
-    id: 34,
-    name: "Bianca Costa",
-    country: "Brazil",
-    isFollowing: false,
-    joinDate: "2022-03-19",
-    contributions: 47,
-    lastActive: "2024-06-12",
-    status: "active",
-    transactionHash: "0x1d2e3f4a5b6c7d8e9f0a1b2c3d4e5f6a7b8c9"
-  },
-  {
-    id: 35,
-    name: "Carlos Mendoza",
-    country: "Mexico",
-    isFollowing: false,
-    joinDate: "2023-06-05",
-    contributions: 35,
-    lastActive: "2024-04-20",
-    status: "inactive",
-    transactionHash: "0x2e3f4a5b6c7d8e9f0a1b2c3d4e5f6a7b8c9d0"
-  },
-  {
-    id: 36,
-    name: "Dimitri Petrov",
-    country: "Russia",
-    avatar: "/avatars/dimitri.jpg",
-    isFollowing: true,
-    joinDate: "2021-08-16",
-    contributions: 97,
-    lastActive: "2024-08-09",
-    status: "active",
-    transactionHash: "0x3f4a5b6c7d8e9f0a1b2c3d4e5f6a7b8c9d0e1"
-  },
-  {
-    id: 37,
-    name: "Emma Thompson",
-    country: "Canada",
-    isFollowing: false,
-    joinDate: "2022-11-22",
-    contributions: 39,
-    lastActive: "2024-07-26",
-    status: "active",
-    transactionHash: "0x4a5b6c7d8e9f0a1b2c3d4e5f6a7b8c9d0e1f2"
-  },
-  {
-    id: 38,
-    name: "Fatima Al-Zahra",
-    country: "Morocco",
-    isFollowing: false,
-    joinDate: "2023-01-31",
-    contributions: 25,
-    lastActive: "2024-03-07",
-    status: "inactive",
-    transactionHash: "0x5b6c7d8e9f0a1b2c3d4e5f6a7b8c9d0e1f2a3"
-  },
-  {
-    id: 39,
-    name: "Giovanni Rossi",
-    country: "Italy",
-    avatar: "/avatars/giovanni.jpg",
-    isFollowing: true,
-    joinDate: "2021-09-11",
-    contributions: 78,
-    lastActive: "2024-08-03",
-    status: "active",
-    transactionHash: "0x6c7d8e9f0a1b2c3d4e5f6a7b8c9d0e1f2a3b4"
-  },
-  {
-    id: 40,
-    name: "Hannah Park",
-    country: "South Korea",
-    isFollowing: false,
-    joinDate: "2022-07-28",
-    contributions: 54,
-    lastActive: "2024-05-24",
-    status: "active",
-    transactionHash: "0x7d8e9f0a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5"
-  },
-  {
-    id: 41,
-    name: "Ivan Volkov",
-    country: "Ukraine",
-    isFollowing: false,
-    joinDate: "2023-04-03",
-    contributions: 18,
-    lastActive: "2024-02-29",
-    status: "inactive",
-    transactionHash: "0x8e9f0a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6"
-  },
-  {
-    id: 42,
-    name: "Jasmine Lee",
-    country: "South Korea",
-    avatar: "/avatars/jasmine.jpg",
-    isFollowing: true,
-    joinDate: "2021-10-17",
-    contributions: 87,
-    lastActive: "2024-08-12",
-    status: "active",
-    transactionHash: "0x9f0a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7"
-  },
-  {
-    id: 43,
-    name: "Klaus Wagner",
-    country: "Austria",
-    isFollowing: false,
-    joinDate: "2022-05-09",
-    contributions: 33,
-    lastActive: "2024-06-03",
-    status: "active",
-    transactionHash: "0x0a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8"
-  },
-  {
-    id: 44,
-    name: "Luca Bernini",
-    country: "Italy",
-    isFollowing: false,
-    joinDate: "2023-12-14",
-    contributions: 9,
-    lastActive: "2024-01-28",
-    status: "inactive",
-    transactionHash: "0x1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
-  },
-  {
-    id: 45,
-    name: "Maya Johansson",
-    country: "Sweden",
-    avatar: "/avatars/maya.jpg",
-    isFollowing: true,
-    joinDate: "2021-07-21",
-    contributions: 112,
-    lastActive: "2024-08-08",
-    status: "active",
-    transactionHash: "0x2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0"
-  },
-  {
-    id: 46,
-    name: "Nicolas Moreau",
-    country: "France",
-    isFollowing: false,
-    joinDate: "2022-01-26",
-    contributions: 61,
-    lastActive: "2024-07-17",
-    status: "active",
-    transactionHash: "0x3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1"
-  },
-  {
-    id: 47,
-    name: "Olga Kozlova",
-    country: "Russia",
-    isFollowing: false,
-    joinDate: "2023-08-02",
-    contributions: 14,
-    lastActive: "2024-04-05",
-    status: "inactive",
-    transactionHash: "0x4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2"
-  },
-  {
-    id: 48,
-    name: "Pablo Hernandez",
-    country: "Spain",
-    avatar: "/avatars/pablo.jpg",
-    isFollowing: true,
-    joinDate: "2021-11-29",
-    contributions: 79,
-    lastActive: "2024-08-07",
-    status: "active",
-    transactionHash: "0x5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3"
-  },
-  {
-    id: 49,
-    name: "Quinn Johnson",
-    country: "Australia",
-    isFollowing: false,
-    joinDate: "2022-09-18",
-    contributions: 44,
-    lastActive: "2024-06-21",
-    status: "active",
-    transactionHash: "0x6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4"
-  },
-  {
-    id: 50,
-    name: "Raj Kumar",
-    country: "India",
-    isFollowing: false,
-    joinDate: "2023-05-30",
-    contributions: 20,
-    lastActive: "2024-03-11",
-    status: "inactive",
-    transactionHash: "0x7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5"
-  },
-  {
-    id: 51,
-    name: "Sarah Williams",
-    country: "United States",
-    avatar: "/avatars/sarah.jpg",
-    isFollowing: true,
-    joinDate: "2021-02-14",
-    contributions: 134,
-    lastActive: "2024-08-12",
-    status: "active",
-    transactionHash: "0x8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6"
-  },
-  {
-    id: 52,
-    name: "Tomas Horvat",
-    country: "Slovakia",
-    isFollowing: false,
-    joinDate: "2022-12-20",
-    contributions: 27,
-    lastActive: "2024-07-09",
-    status: "active",
-    transactionHash: "0x9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7"
-  },
-  {
-    id: 53,
-    name: "Uma Krishnan",
-    country: "India",
-    isFollowing: false,
-    joinDate: "2023-07-18",
-    contributions: 16,
-    lastActive: "2024-05-02",
-    status: "inactive",
-    transactionHash: "0x0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8"
-  },
-  {
-    id: 54,
-    name: "Viktor Jansen",
-    country: "Netherlands",
-    avatar: "/avatars/viktor.jpg",
-    isFollowing: true,
-    joinDate: "2021-04-26",
-    contributions: 101,
-    lastActive: "2024-08-05",
-    status: "active",
-    transactionHash: "0x1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9"
-  },
-  {
-    id: 55,
-    name: "Wei Zhang",
-    country: "China",
-    isFollowing: false,
-    joinDate: "2022-08-12",
-    contributions: 50,
-    lastActive: "2024-06-30",
-    status: "active",
-    transactionHash: "0x2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0"
-  },
-  {
-    id: 56,
-    name: "Ximena Rodriguez",
-    country: "Colombia",
-    isFollowing: false,
-    joinDate: "2023-02-21",
-    contributions: 26,
-    lastActive: "2024-04-17",
-    status: "inactive",
-    transactionHash: "0x3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1"
-  },
-  {
-    id: 57,
-    name: "Yasmin Hassan",
-    country: "Egypt",
-    avatar: "/avatars/yasmin.jpg",
-    isFollowing: true,
-    joinDate: "2021-09-08",
-    contributions: 93,
-    lastActive: "2024-08-10",
-    status: "active",
-    transactionHash: "0x4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2"
-  },
-  {
-    id: 58,
-    name: "Zoe Mitchell",
-    country: "New Zealand",
-    isFollowing: false,
-    joinDate: "2022-10-31",
-    contributions: 37,
-    lastActive: "2024-07-05",
-    status: "active",
-    transactionHash: "0x5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2c3"
-  },
-  {
-    id: 59,
-    name: "Ahmed Ben Ali",
-    country: "Tunisia",
-    isFollowing: false,
-    joinDate: "2023-09-12",
-    contributions: 11,
-    lastActive: "2024-02-18",
-    status: "inactive",
-    transactionHash: "0x6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2c3d4"
-  },
-  {
-    id: 60,
-    name: "Bethany Davis",
-    country: "Canada",
-    avatar: "/avatars/bethany.jpg",
-    isFollowing: true,
-    joinDate: "2021-01-08",
-    contributions: 147,
-    lastActive: "2024-08-11",
-    status: "active",
-    transactionHash: "0x7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2c3d4e5"
-  }
-];
